+++ src/main/java/com/around/wmmarket/controller/DealReviewApiController.java
@@ -28,10 +28,10 @@ public class DealReviewApiController {
         // check signedUser
         if(signedUser==null) return ResponseEntity.badRequest().body("login 을 먼저 해주세요");
         DealPost dealPost=dealPostService.getDealPost(requestDto.getDealPostId());
-        if(signedUser.getUsername().equals(dealPost.getUser().getEmail())) return ResponseEntity.badRequest().body("판매자는 본인글의 후기를 남길 수 없습니다.");
+        if(dealPost.getUser()==null||signedUser.getUsername().equals(dealPost.getUser().getEmail())) return ResponseEntity.badRequest().body("판매자는 본인글의 후기를 남길 수 없습니다.");
         // check dealState && dealSuccess
         if(dealPost.getDealState()!=DealState.DONE) return ResponseEntity.badRequest().body("게시글 상태가 DONE 이 아닙니다. 거래상태코드:"+dealPost.getDealState());
-        if(!dealPost.getDealSuccess().getBuyer().getEmail().equals(signedUser.getUsername())) return ResponseEntity.badRequest().body("거래글의 구매자와 로그인 유저가 일치하지 않습니다. buyerEmail:"+dealPost.getDealSuccess().getBuyer().getEmail());
+        if(dealPost.getDealSuccess().getBuyer()==null||!dealPost.getDealSuccess().getBuyer().getEmail().equals(signedUser.getUsername())) return ResponseEntity.badRequest().body("거래글의 구매자와 로그인 유저가 일치하지 않습니다.");
 
         dealReviewService.save(signedUser.getUsername(),requestDto.getContent(),dealPost);
 
@@ -49,7 +49,7 @@ public class DealReviewApiController {
         if(signedUser==null) return ResponseEntity.badRequest().body("login 을 먼저 해주세요");
         DealReview dealReview=dealReviewRepository.findById(requestDto.getDealReviewId())
                 .orElseThrow(()->new NoSuchElementException("해당 리뷰글이 없습니다. reviewId:"+requestDto.getDealReviewId()));
-        if(!dealReview.getBuyer().getEmail().equals(signedUser.getUsername())) return ResponseEntity.badRequest().body("리뷰 작성자가 아닙니다.");
+        if(dealReview.getBuyer()==null||!dealReview.getBuyer().getEmail().equals(signedUser.getUsername())) return ResponseEntity.badRequest().body("리뷰 작성자가 아닙니다.");
         // update
         dealReviewService.update(requestDto);
         return ResponseEntity.ok().body("update success");
@@ -61,7 +61,7 @@ public class DealReviewApiController {
         if(signedUser==null) return ResponseEntity.badRequest().body("login 을 먼저 해주세요");
         DealReview dealReview=dealReviewRepository.findById(dealReviewId)
                 .orElseThrow(()->new NoSuchElementException("해당 리뷰글이 없습니다. reviewId:"+dealReviewId));
-        if(!dealReview.getBuyer().getEmail().equals(signedUser.getUsername())) return ResponseEntity.badRequest().body("리뷰 작성자가 아닙니다.");
+        if(dealReview.getBuyer()==null||!dealReview.getBuyer().getEmail().equals(signedUser.getUsername())) return ResponseEntity.badRequest().body("리뷰 작성자가 아닙니다.");
         // delete
         dealReviewService.delete(dealReviewId);
         return ResponseEntity.ok().body("delete success");
